fix: debugging login + missing bridge methods
